+++ src/renderable/container.js
@@ -653,11 +653,12 @@ class Container extends Renderable {
             }
 
             if (!keepalive) {
-                if (typeof (child.destroy) === "function") {
+                // only push back in the pool if implementing "onResetEvent";
+                if (typeof child.onResetEvent === "function") {
+                    pool.push(child);
+                } else if (typeof child.destroy === "function") {
                     child.destroy();
                 }
-
-                pool.push(child);
             }
 
             // Don't cache the child index; another element might have been removed